requirements.txt
aniso8601==7.0.0
annotated-types==0.7.0
anyio==3.7.1
asgiref==3.4.1
blinker==1.5
Brlapi==0.8.4
Brotli==1.1.0
certifi==2023.7.22
chardet==5.1.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==38.0.4
cupshelpers==1.0
dbus-python==1.3.2
defusedxml==0.7.1
diff-match-patch==20230430
distlib==0.3.9
distro==1.8.0
distro-info==1.5+deb12u1
dj-database-url==2.0.0
Django==3.2.20
django-ckeditor==6.0.0
django-crispy-forms==1.13.0
django-import-export==2.9.0
django-jet-4==0.0.1
django-js-asset==1.2.2
django-shortuuidfield==0.1.3
django-taggit==3.0.0
dnspython==2.6.1
email-validator==2.0.0.post2
et-xmlfile==1.1.0
exceptiongroup==1.1.2
fail2ban==1.0.2
filelock==3.16.1
gunicorn==20.1.0
h11==0.14.0
httpcore==0.17.3
httplib2==0.20.4
httpx==0.24.1
idna==3.4
iotop==0.6
lazr.restfulclient==0.14.5
lazr.uri==1.0.6
louis==3.24.0
Mako==1.2.4.dev0
MarkupSafe==2.1.2
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
Pillow==8.4.0
platformdirs==4.3.6
promise==2.3
psycopg2==2.9.9
psycopg2-binary==2.9.6
pycairo==1.20.1
pycups==2.0.1
pycurl==7.45.2
pydantic==1.10.12
pydantic_core==2.18.2
PyGObject==3.42.2
pyinotify==0.9.6
PyJWT==2.8.0
pyparsing==3.0.9
PySimpleSOAP==1.16.2
pysmbc==1.0.23
python-apt==2.6.0
python-dateutil==2.8.2
python-debian==0.1.49
python-debianbts==4.0.1
python-decouple==3.8
python-dotenv==1.0.1
pytz==2022.7.1
pyxdg==0.28
PyYAML==6.0.1
reportbug==12.0.0
requests==2.32.2
Rx==1.6.1
shortuuid==1.0.13
singledispatch==3.6.2
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
systemd-python==235
tablib==3.6.1
text-unidecode==1.3
typing_extensions==4.7.1
ufw==0.36.2
urllib3==2.2.1
wadllib==1.3.6
whitenoise==5.3.0
xdg==5
xlrd==2.0.1
xlwt==1.3.0
